client/requirements.txt
mysql-connector
psycopg2
redis
pymongo